requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
anyjson==0.3.3            # via -r requirements/base.in
boto==2.49.0              # via -r requirements/base.in
certifi==2020.12.5        # via requests
chardet==4.0.0            # via requests
colorama==0.4.4           # via rednose
coverage==5.3.1           # via nosexcover
django-nose==1.4.7        # via -r requirements/base.in
django==2.2.17            # via -r requirements/base.in
dogstatsd-python==0.5.6   # via -r requirements/base.in
gunicorn==0.17.4          # via -r requirements/base.in
idna==2.10                # via requests
lxml==2.3.6               # via -r requirements/base.in
newrelic==2.46.0.37       # via -r requirements/base.in
nose==1.3.7               # via django-nose, nosexcover
nosexcover==1.0.11        # via -r requirements/base.in
numpy==1.6.2              # via -r requirements/base.in
path.py==12.5.0           # via -r requirements/base.in
path==15.0.1              # via path.py
pika==1.1.0               # via -r requirements/base.in
pytz==2020.5              # via django
rednose==1.3.0            # via -r requirements/base.in
requests==2.25.1          # via -r requirements/base.in
south==1.0.2              # via -r requirements/base.in
sqlparse==0.4.1           # via django
termstyle==0.1.11         # via rednose
urllib3==1.26.2           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
